+++ hal/include/HAL/DIO.h
@@ -1,0 +1,29 @@
+#pragma once
+#include <stdint.h>
+
+extern "C" {
+void* initializeDigitalPort(void* port_pointer, int32_t* status);
+void freeDigitalPort(void* digital_port_pointer);
+
+void* allocatePWM(int32_t* status);
+void freePWM(void* pwmGenerator, int32_t* status);
+void setPWMRate(double rate, int32_t* status);
+void setPWMDutyCycle(void* pwmGenerator, double dutyCycle, int32_t* status);
+void setPWMOutputChannel(void* pwmGenerator, uint32_t pin, int32_t* status);
+
+bool allocateDIO(void* digital_port_pointer, bool input, int32_t* status);
+void freeDIO(void* digital_port_pointer, int32_t* status);
+void setDIO(void* digital_port_pointer, short value, int32_t* status);
+bool getDIO(void* digital_port_pointer, int32_t* status);
+bool getDIODirection(void* digital_port_pointer, int32_t* status);
+void pulse(void* digital_port_pointer, double pulseLength, int32_t* status);
+bool isPulsing(void* digital_port_pointer, int32_t* status);
+bool isAnyPulsing(int32_t* status);
+
+void setFilterSelect(void* digital_port_pointer, int filter_index,
+                     int32_t* status);
+int getFilterSelect(void* digital_port_pointer, int32_t* status);
+
+void setFilterPeriod(int filter_index, uint32_t value, int32_t* status);
+uint32_t getFilterPeriod(int filter_index, int32_t* status);
+}
