+++ hal/include/HAL/HAL.h
@@ -9,14 +9,24 @@
 #include <cmath>
 
 #include "Accelerometer.h"
-#include "Analog.h"
+#include "AnalogAccumulator.h"
+#include "AnalogInput.h"
+#include "AnalogOutput.h"
+#include "AnalogTrigger.h"
 #include "Compressor.h"
-#include "Digital.h"
+#include "Counter.h"
+#include "DIO.h"
+#include "Encoder.h"
 #include "Errors.h"
+#include "I2C.h"
 #include "Interrupts.h"
 #include "Notifier.h"
 #include "PDP.h"
+#include "PWM.h"
 #include "Power.h"
+#include "Relay.h"
+#include "SPI.h"
+#include "SPIAccumulator.h"
 #include "SerialPort.h"
 #include "Solenoid.h"
