+++ hal/include/HAL/PWM.h
@@ -1,0 +1,15 @@
+#pragma once
+#include <stdint.h>
+
+extern "C" {
+bool checkPWMChannel(void* digital_port_pointer);
+
+void setPWM(void* digital_port_pointer, unsigned short value, int32_t* status);
+bool allocatePWMChannel(void* digital_port_pointer, int32_t* status);
+void freePWMChannel(void* digital_port_pointer, int32_t* status);
+unsigned short getPWM(void* digital_port_pointer, int32_t* status);
+void latchPWMZero(void* digital_port_pointer, int32_t* status);
+void setPWMPeriodScale(void* digital_port_pointer, uint32_t squelchMask,
+                       int32_t* status);
+uint16_t getLoopTiming(int32_t* status);
+}
