+++ wpilibc/athena/src/I2C.cpp
@@ -6,7 +6,6 @@
 /*----------------------------------------------------------------------------*/
 
 #include "I2C.h"
-#include "HAL/Digital.h"
 #include "HAL/HAL.h"
 #include "WPIErrors.h"
